-Added UI for displaying node cost and disallowed travel directions. -TODO: bug where shortestPath ignores disallowed travel directions.

+++ Assets/Scripts/TradingMiniGame/Controller/Implementation/CustomGridModifier.cs
@@ -10,7 +10,7 @@ namespace TradingMiniGame
     public class CustomGridModifier : MonoBehaviour, IGameGridSetupModifier
     {
         private IGameGridController _gameGridController;
-        public float maxPathCost;
+        public int maxPathCost;
         public int maxNeighborsRemoved;
 
         [Inject]
